fix: Fix update-repo ready for further testing
We now always clone repos fully instead of with depth=1 - the latter
breaks anything that uses the commit log. (We can potentially add
a flag for that later.)

The change to filtering in GetApiCommits mirrors what we do in .NET;
this is a much, much faster way of finding the commits that actually
change directories. We don't even try to handle multi-parent commits,
but I don't think that's required at the moment.