feat(generator/rust)!: use integers for enums (#1448)
Change the enums to be integer-based as opposed to string-based. This
changes the API, so it is a breaking change. We need to make this change
to support client libraries based on gRPC. Unknown enums are received
as integers in that case. We need to make the change for all client
libraries, as any of them may need to switch the underlying transport if
they gain streaming RPCs.