feat: Allow git user name/email to be passed in by flags
At the same time, remove the separate function for committing
everything in a release PR. There were TODOs there, but if we fail to
commit a change in a release PR, that should be fatal anyway - it's
not like the "this library is broken, so skip it" case.

Additionally, when committing, don't show the repo status. (That was
basically useful when debugging, but isn't as useful now.)