feat: Rework the configure command ready for automation
This still allows an individual API to be configured, but will usually
be called in a Kokoro job to configure "all APIs that look like they
need configuring". The resulting PR does not include generated code;
we expect generation to occur on the next regular update-apis cycle.
However, the library is generated and built so that errors can be
reported and investigated early. (Any failure to generate/build will
result in the PR reporting the failure but not including the API.)