feat: Support library-id flag in create-release-pr
This restricts the command to only releasing the specified library,
and causes the command to fail if that library *isn't* then released
(either because there are no changes, or potentially because there's a
typo in the library name, or because the integration tests failed; at
the moment it's up to the user to figure out which).

Currently this does *not* force a library to be released if it
wouldn't otherwise be. We could potentially add that functionality if
it proves desirable. (Otherwise users will have to create "no-op"
commits that affect a path specified for the library.)