feat: Overhaul how pipeline state and config are loaded
This introduces an extra bit of configurability for commands: once
they've opened/cloned/ignored the language repo, RunCommand asks them
to (maybe) load pipeline state and config from wherever they wish to.
This is then used to work out which image to use, and configuration
for container commands.

Most commands do this from the language repo they've just
cloned/opened, but publish-release-artifacts does it from files under
the "artifact root" passed in on the command line. (These are now
copied during create-release-artifacts.) This fixes b/416449513

Additionally, the generate command now determines whether to use a
local repo or a clone very early on - which means its pipeline state
and config can be used accordingly.

The implementation of this required some more code to load/save
the files, so that's now in a separate Go source file
(stateandconfig.go).