feat(internal/cli): add Command.Usage and Command.Long fields
The librarian commands were previously undocumented. Added Usage and
Long fields to the Command struct to serve as placeholders for usage
text.  For now, these fields are populated with a TODO as a
placeholders.

When constructing the usage information, the code will now panic if
Short, Long, or Usage are empty, so that we ensure new commands are
documented moving forward.

For https://github.com/googleapis/librarian/issues/237