feat(internal/cli,librarian): use cli framework for Librarian
The CLI framework now supports subcommands and is used to construct the
top-level Librarian command. This aligns its construction with other
commands and makes it possible for other subcommands to be supported in
the future.

Fixes https://github.com/googleapis/librarian/issues/469

---

This commit is based off of
https://github.com/googleapis/librarian/pull/478. It is ready for
review, and will be rebased once
https://github.com/googleapis/librarian/pull/478 is merged. This section
of the commit message will then be removed.