fix(internal/librarian): don't display both usage twice
If Command.Parse fails having found a specific command, we can expect
that command to have displayed a suitable command-specific-usage
message; we don't need to display the generic usage as well.

This affects both the case of "help explicitly requested with -h" as
well as "a flag that's not registed for this command was supplied".

Fixes #666