chore: remove pull_request_review trigger from multi_approvers.yaml
When a PR is initiated from a fork, GitHub actions that are triggered based on `pull_request_review` do not receive the needed secrets. So, multi-approvers can't work. Additionally, for non-fork PRs, multi-approvers will already be green anyway because all users with permission to create branches are Googlers. So, the `pull_request_review` trigger is useless.

Signed-off-by: Mike Eltsufin <meltsufin@google.com>