refactor(internal/librarian): remove contentLoader (#1698)
The release code will also need to use some of this functionality and to
make the code easier to call/read I refactored out the idea of a content
loader that I believe was added to help coverage. I am fine taking a
couple points of a coverage hit to make the code, in my opinion, more
readable.

Updates: #1009