feat(internal/conventionalcommits): support commit override and nested commits in parsing (#1776)
This change moves the parsing logic to a dedicated
internal/conventionalcommits package and adds support for commit
overrides and nested commits.

- Commit Override: A BEGIN_COMMIT_OVERRIDE/END_COMMIT_OVERRIDE block can
be used to replace the entire commit message with the content inside the
block.
- Nested Commits: BEGIN_NESTED_COMMIT/END_NESTED_COMMIT blocks allow for
multiple conventional commits to be parsed from a single commit message.

Note: an "isNested" field is added for requirement "nested commits
information will only be used for release notes, and not for bumping
semver", see #1769

refer to
[go/librarian:release-please-lite](http://go/librarian:release-please-lite#heading=h.5bformyuaenc)
for discussions on supported/unsupported edge case scenarios.

For #1695