tests(internal): avoid a panic when expecting an error (#2217)
These changes all use Fatal or Fatalf to abort a test when an expected
error isn't returned (i.e. `err == nil`). This avoids a panic when
further checks are performed on the error, typically checking its
message.

Some tests used if/else-if to avoid this panic. This change updates all
tests to a consistent style.

For internal/sidekick, this only changes the "definitely wrong" cases.
There are some other tests which are currently fine due to only checking
whether or not there *is* an error, but which would need modification if
the test were changed to check for a particular error message.

Some other tests are

Fixes #2216