feat(sidekick): parse discovery doc path templates
Discovery doc path templates are specified using URI templates, that is
RFC 6570 templates. This is not the same format as the HTTP annotations
for Protobuf.  Compute only uses a **subset** of the URI template spec,
which allows us to parse the strings and use (without change) the
`api.PathTemplate` structure.

If compute starts using more complex path templates we can grow the
data structures at that time.