fix(librarian): temporarily use short SHA for release links
We noticed that we were not able to open the release init PR today
because the character limit exceeds what github allows. For now, by
shortening the SHA links we can get under this limit. We should have
a more robust solution in the future.

Related-disscussion: https://github.com/orgs/community/discussions/27190
Related: #2234
Fixes: #2321