fix(internal/librarian): avoid work duplication when finding changes
This changes how we format a generation PR. Instead of finding which
files have changed in the language repo (an operation which doesn't
depend on the library) once for each library, we do so once at the
start. We can then skip libraries which haven't changed, without even
looking at the API source repo.

Fixes #2557