refactor(librarian): ParseCommit now returns a single commit
The main change here is that ParseCommit now returns a single commit
that is a one to one mapping with the orginal git commit. It does
this by adding the idea of NestedCommits that are associated this
a parent commit. This can help up better keep track of related commits.

In future refactorings it would be nice to have this package just return
ConventionalCommits and not the wrapped git Commit object.

Updates: #2543