fix(librarian): truncate pr body info it release notes too big
This will reattempt to create a new body if the first attempt was
too big. If the body is still to big we will still return an error.
I tested this in a side program and this should get us up to 350+
libraries which seems like enough for now at least.

Fixes: #2543