feat: speed up finding changed files in commits
We don't need the actual diffs between commits - only the names of the
changed files. This massively improves performance of "release init".

While there *may* be other things we want to do in the future, this
makes enough of a difference that we can probably hold off on larger
changes (at least for the sake of performance).

Fixes #2634