fix: fix the reference to the old "librariangen" folder
The cloudbuild-exitgate.yaml was referencing the old folder
librariangen.

This should resolve Cloud Build failure in b/454950041:

```
unable to prepare context: unable to evaluate symlinks in Dockerfile path: lstat /workspace/internal/librariangen/Dockerfile: no such file or directory
"gcr.io/cloud-builders/docker" failed: step exited with non-zero status: 1
```

This change also updates the reference to "librariangen" as it's just the binary name of Java and Go language containers.
The term "librariangen" should not be used in a package to be shared with other language containers.