feat(internal/container/java): generate pom.xml files for generated libraries
Add creation of `pom.xml` files for the generated Java client libraries.
We assume that the `google-cloud-pom-parent` and `google-cloud-jar-parent` are published from a separate repository. For now, the version is hard-coded in the templates.

`run-genrate-library.sh` is also updated to not only generate the sample library, but to build it as well.

Fixes #2657