feat(surfer): implement surfer generate command
Create the initial surfer CLI and add the framework for `surfer
generate`. The generation logic will be added in a follow-up change.

Move `testdata/gcloud.yaml` to `testdata/parallelstore/gcloud.yaml`
to make it clear that it belongs to the parallelstore service.

Remove gcloud.Generate from sidekick, since we no longer intend to run
it from the sidekick CLI.