feat(librarian): set a default version for repo metadata
A new parameter is added to repometadata.GenerateRepoMetadata to set
the default version. The longer term plan is still to compute this,
but until that's ready we will allow the caller to specify it.
Callers who want to generate repo metadata without a default version
can pass in an empty string.