feat(internal/python): introduce Python generation
The language "python" can now be used to generate Python libraries.
This assumes a complete working environment for the generator, which
requires:

- protoc
- python3
- The Python GAPIC generator
- synthtool

This includes a wrapper around exec.CommandContext for the purposes of
mocking. After further refinement, this can be moved into the command
package. See #3206 for more details.

Towards #3107.