fix: fix compilation error in generate.go
Possibly due to a bad merge, we ended up with return statements in
generate which didn't match the signature.

In fixing these, we need to make sure that the `err` variable used in
anything that doesn't introduce a new scope still refers to the
variable introduced at the top level of the function.

Fixes #3212