refactor(internal/librarian): replace test mocks with fake language
The testhelper language is renamed to fake. fakeCreate, fakeFormat, and
fakePublish functions are added to support testing for all commands.

The fake language also now uses channels and service configs like rust
and python, ensuring it exercises the same code paths.

TestCreateLibrary and TestPublish are rewritten to use the fake language
instead of mockGenerator and mockRustHelper. The tests now call the
actual create and publish functions and verify the files they produce.
This also removes logic for mock setup code.

createLibrarianYaml and validateLibrarianYaml are inlined, since they
are now each only called in one place. TestAddLibraryToLibrarianYaml now
uses cmp.Diff to verify channels, replacing nested if-else statements
with explicit test expectations

Language strings are replaced with constants throughout the codebase.

A bug in deriveOutput is fixed where the output parameter was ignored for
non-rust languages.